move ansi from data to resources
